+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Validate CHANGELOG in CI workflow
 - Close CHANGELOG version automatically when a new release is made
-- Refactor code to adds an optional stream parameter in the run method of Process 
+- Handle input for `Process`es
 
 ### Changed
 - Update `specs2` to 4.9.3